requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
supabase>=2.18.0
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-tavily>=0.1.0
pydantic>=2.5.0
streamlit>=1.28.0
python-dotenv>=1.0.0